Remove code generation package from the project
The scaffolding tools I wanted to use require a database context
for the model they are to build the scaffold for. Since this
project is not intended to need a database then I won't be able to
use these scaffolding tools.

Instead I will look at the existing Razor pages to see if they can
be repurposed to suit the app needs. Essentially this will just be
a single page web site that displays a generated maze from simple
inputs.