Add interface for accessing grid cells
The `Grid` class abstracts a collection of `Cell` objects. The grid
access is made with X and Y coordinates where the first set of
coordinates 1, 1 starts at the top left of the grid. If a zero
based index is required the conversion of the coordinates is not
done until the last possible moment.