Display a grid at the view as ASCII art
An automatic code formatter changed all the white space in Grid.cs
making it look like the entire file changed. The `ToString()`
method and its helpers are the important change, allowing the grid
to be represented as a string.

The variables `i` and `j` used in `for` loops have been replaced
with `x` and `y` as appropriate to aid in reading the code as a
grid of cells acessed with x and y coordinates.