Create mazes with Sidewinder algorithm
This change includes a modification to the `Grid` class iterator.
Originally it was intended that foreach loops would iterate
across the X axis of the grid first and then down the Y axis i.e.
moving row by row. The sidewinder algorithm was implemented to
expect this order of iteration across the grid. As it was before
the change the iterator was moving down the Y axis first.