Extract `Grid` PNG writing into separate class
Writing the PNG is a substantial task. The `Grid` class is about
how a maze is internally represented so it is reasoned that writing
an image file should be the responsibility of another class.

Access to the new `MazeImage` class is kept internal as it is only
intended to be used as an aid to producing PNG files of `Grid`
instances.